1. HDR with Luminosity 2. Download with 'az copy'

+++ storage.py
@@ -2,9 +2,12 @@
 import os
 from pathlib import Path
 import shutil
+import subprocess
 from azure.storage.blob import BlobServiceClient, BlobClient, ContainerClient
 import settings
 import time
+
+from utils import run_command
 
 # From: https://stackoverflow.com/questions/63413832/upload-local-folder-to-azure-blob-storage-using-blobserviceclient-with-python-v1
 def _upload_file(client, source, dest):
@@ -44,6 +47,13 @@ def _get_connection_string():
     return connect_str
 
 
+def _get_sas_key():
+    key = os.getenv('TIMELAPSE_AZURE_STORAGE_SAS_KEY')
+    if not key:
+        raise Exception('Environment variable TIMELAPSE_AZURE_STORAGE_SAS_KEY is not set. Get it from Azure Portal')
+    return key 
+
+
 def upload_to_remote_storage(container_name, source, dest, delete):
     connect_str = _get_connection_string()
     service_client = BlobServiceClient.from_connection_string(connect_str)
@@ -57,19 +67,46 @@ def get_latest_files(container_name, day):
     client = service_client.get_container_client(container_name)
     if day is not None:
         start_with = f'{settings.TIMELAPSE_NAME}/images/{day}'
+    last_dl = None
     while True:
         blobs = client.list_blobs(name_starts_with=start_with)
         blobs = list(blobs)
         blobs.sort(key=lambda b: b.name)
 
         print(blobs[-1].name)
-
-        download_file_path = '/tmp/dl.jpg'
-        with open(download_file_path, "wb") as download_file:
-            dl = client.download_blob(blobs[-1]).readall()
-            download_file.write(dl)
+        if last_dl != blobs[-1].name:
+            download_file_path = '/tmp/dl.jpg'
+            with open(download_file_path, "wb") as download_file:
+                dl = client.download_blob(blobs[-1]).readall()
+                download_file.write(dl)
+                last_dl = blobs[-1].name
         time.sleep(1)
 
 
-    
-    
+def download_files_DEPRECATED(container_name, date_prefix, dst_dir):
+    connect_str = _get_connection_string()
+    service_client = BlobServiceClient.from_connection_string(connect_str)
+    client = service_client.get_container_client(container_name)
+    start_with = f'{settings.TIMELAPSE_NAME}/images/{date_prefix}'
+    blobs = client.list_blobs(name_starts_with=start_with)
+    blobs = list(blobs)
+    blobs.sort(key=lambda b: b.name)
+    for blob in blobs:
+        download_file_path = dst_dir / blob.name
+
+        if ~download_file_path.exists():
+            download_file_path.parent.mkdir(parents=True, exist_ok=True)
+            with open(download_file_path, "wb") as download_file:
+                dl = client.download_blob(blob).readall()
+                download_file.write(dl)
+
+
+def download_files(container_name: str, date_prefix: str, dst_dir: Path, overwrite: bool):
+    sas_key = _get_sas_key()
+    in_dir = f'{settings.TIMELAPSE_NAME}/images'
+    dst_dir2 = f'{dst_dir}/{in_dir}'
+    Path(dst_dir2).mkdir(parents=True, exist_ok=True)
+    url = f'https://annolapse.blob.core.windows.net/{container_name}/{in_dir}/*'
+    overwrite_str = 'true' if overwrite else 'false'
+    cmd = rf'azcopy cp "{url}?{sas_key}" "{dst_dir2}" --overwrite={overwrite_str} --from-to=BlobLocal --recursive --include-pattern "{date_prefix}*"'
+    run_command(cmd)
